Create AYRUser class and refactor business logic that depends on ayr group logic
