+++ app/tests/test_browse.py
@@ -317,7 +317,7 @@ def test_browse_transferring_body(client: FlaskClient):
     file = files[0]
     transferring_body_id = file.file_consignments.consignment_bodies.BodyId
 
-    mock_standard_user(client, file.file_consignments.consignment_bodies.Name)
+    mock_standard_user(client, [file.file_consignments.consignment_bodies.Name])
 
     response = client.get(
         f"/browse?transferring_body_id={transferring_body_id}"
@@ -365,7 +365,7 @@ def test_browse_series(client: FlaskClient):
     file = files[0]
     series_id = file.file_consignments.SeriesId
 
-    mock_standard_user(client, file.file_consignments.consignment_bodies.Name)
+    mock_standard_user(client, [file.file_consignments.consignment_bodies.Name])
 
     response = client.get(f"/browse?series_id={series_id}")
 
@@ -413,7 +413,7 @@ def test_browse_consignment(client: FlaskClient):
     file = files[0]
     consignment_id = file.file_consignments.ConsignmentId
 
-    mock_standard_user(client, file.file_consignments.consignment_bodies.Name)
+    mock_standard_user(client, [file.file_consignments.consignment_bodies.Name])
 
     response = client.get(f"/browse?consignment_id={consignment_id}")
 
@@ -460,7 +460,7 @@ def test_browse_consignment_with_missing_file_metadata(client: FlaskClient):
     file = files[10]
     consignment_id = file.file_consignments.ConsignmentId
 
-    mock_standard_user(client, file.file_consignments.consignment_bodies.Name)
+    mock_standard_user(client, [file.file_consignments.consignment_bodies.Name])
 
     response = client.get(f"/browse?consignment_id={consignment_id}")
 
@@ -511,7 +511,7 @@ def test_browse_consignment_filter_display_multiple_pages(
 
     create_multiple_files_for_consignment(consignment_id)
 
-    mock_standard_user(client, file.file_consignments.consignment_bodies.Name)
+    mock_standard_user(client, [file.file_consignments.consignment_bodies.Name])
 
     response = client.get(f"/browse?page=2&consignment_id={consignment_id}")
     assert response.status_code == 200
